Split polling __gtg into its own task
