+++ bin/next-build-tools.js
@@ -6,6 +6,7 @@ require('isomorphic-fetch');
 
 var program = require('commander');
 var deploy = require('../tasks/deploy');
+var waitForGtg = require('../tasks/wait-for-gtg');
 var configure = require('../tasks/configure');
 var scale = require('../tasks/scale');
 var provision = require('../tasks/provision');
@@ -231,6 +232,15 @@ program
 	.action(function(apps) {
 		return rebuild({
 			apps: apps
+		}).catch(exit);
+	});
+
+program
+	.command('wait-for-gtg <app>')
+	.description('Polls the /__gtg endpoint of a given app until it returns 200')
+	.action(function(app) {
+		return waitForGtg({
+			app: app
 		}).catch(exit);
 	});
